tests/wpt/web-platform-tests/old-tests/submission/Opera/script_scheduling/096.html

<!DOCTYPE html>
<html><head>
	<title> scheduler: defer script added from document.write relative to DOMContentLoaded</title>
	<script src="/resources/testharness.js"></script>
	<script src="/resources/testharnessreport.js"></script>
	<script src="testlib/testlib.js"></script>
</head>
<body>

	<div id="log">FAILED (This TC requires JavaScript enabled)</div>

        <script>
          log("inline script #1");
          document.write("<script defer src='scripts/include-1.js'><\/script>")
        </script>
	<script>
          log("inline script #2");
          var t = async_test();

          addEventListener("DOMContentLoaded", t.step_func(function() {assert_array_equals(eventOrder, ["inline script #1", "inline script #2"])}), false);

          onload = t.step_func(function() {assert_array_equals(eventOrder, ["inline script #1", "inline script #2", "external script #1"]); t.done();});
        </script>
</body></html>
